refactoring code to function declaration rather then function expressions. also refactoring the code to external file and initializing the game from index.js
